Commenting!
Commenting has now been prototyped! You can add comments to each of the tiles, and all comments for that tile are displayed. These will also differentiate lecturers and students, currently just the colour of the icon.
We will need to add dates but this should be manageable once we have a column for it.