Drag & Dropping - Phase 1
That's right gents, you can now drag and drop the components when editing the tiles! And no, I haven't tested it in light theme.
There is a bug with new components in that they aren't actually given a proper order value yet, but that's exactly what I'm going onto next.