Unit Editing Complete
Lecturers can now edit the Unit Code, Name, and Description from the Class Info Modal. This updates in the database and on the client-side too!

Removed unnecessary login file and cleaned up some code/comments.

NOTE: A new session variable was added, so if you're still logged in when pulling this commit, please log out and back in to avoid errors.